+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
 
+### v1.2 - 27/02/2019
+
+Special thanks to [@kitzberger](https://github.com/kitzberger)
+
+- add setPreset and playPreset
+- some improvements
+
+
 ### v1.1 - 01/02/2019
 
 - Commandes supplémentaires pour le contrôle dans Jeedom